+++ src/components/calcite-dropdown-group/calcite-dropdown-group.tsx
@@ -8,7 +8,6 @@ import {
   Listen,
   Prop
 } from "@stencil/core";
-import { getElementTheme } from "../../utils/dom";
 import { guid } from "../../utils/guid";
 
 @Component({
@@ -73,14 +72,12 @@ export class CalciteDropdownGroup {
   }
 
   render() {
-    const theme = getElementTheme(this.el);
-    const scale = this.el.closest('calcite-dropdown')?.scale || this.el.assignedSlot.closest('calcite-dropdown').scale || 'm'
     const groupTitle = this.groupTitle ? (
       <span class="dropdown-title">{this.groupTitle}</span>
     ) : null;
 
     return (
-      <Host theme={theme} scale={scale}>
+      <Host>
         {groupTitle}
         <slot />
       </Host>