+++ src/components/calcite-modal/calcite-modal.tsx
@@ -10,7 +10,6 @@ import {
   Method,
   State
 } from "@stencil/core";
-import { x24 } from "@esri/calcite-ui-icons";
 import { queryShadowRoot, isHidden, isFocusable } from "@a11y/focus-trap";
 import { getElementDir, getElementTheme } from "../../utils/dom";
 
@@ -83,15 +82,7 @@ export class CalciteModal {
               ref={el => (this.closeButton = el)}
               onClick={() => this.close()}
             >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                height="24"
-                width="24"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-              >
-                <path d={x24} />
-              </svg>
+              <calcite-icon icon="x" scale="m"></calcite-icon>
             </button>
             <header class="modal__title">
               <slot name="header" />
@@ -99,7 +90,7 @@ export class CalciteModal {
           </div>
           <div
             class={{
-              "modal__content": true,
+              modal__content: true,
               "modal__content--spaced": !this.noPadding
             }}
             ref={el => (this.modalContent = el)}
@@ -188,11 +179,7 @@ export class CalciteModal {
       el.focus();
       return;
     }
-    const focusableElements = queryShadowRoot(
-      this.el,
-      isHidden,
-      isFocusable
-    );
+    const focusableElements = queryShadowRoot(this.el, isHidden, isFocusable);
     if (focusableElements.length > 0) {
       focusableElements[0].focus();
     } else {