+++ src/components/calcite-popover/calcite-popover.tsx
@@ -18,11 +18,11 @@ import {
   updatePopper
 } from "../../utils/popper";
 import { Modifier, Placement, Instance as Popper } from "@popperjs/core";
-import { VNode } from "@stencil/state-tunnel/dist/types/stencil.core";
-import { x16 } from "@esri/calcite-ui-icons";
-import CalciteIcon from "../../utils/CalciteIcon";
+import { VNode } from "@stencil/core/internal/stencil-core";
 import { guid } from "../../utils/guid";
 import { HOST_CSS } from "../../utils/dom";
+
+type FocusId = "close-button";
 
 /**
  * @slot image - A slot for adding an image. The image will appear above the other slot content.
@@ -151,6 +151,8 @@ export class CalcitePopover {
 
   arrowEl: HTMLDivElement;
 
+  closeButtonEl: HTMLButtonElement;
+
   // --------------------------------------------------------------------------
   //
   //  Lifecycle
@@ -196,6 +198,16 @@ export class CalcitePopover {
           popper
         })
       : this.createPopper();
+  }
+
+  @Method()
+  async setFocus(focusId?: FocusId) {
+    if (focusId === "close-button") {
+      this.closeButtonEl?.focus();
+      return;
+    }
+
+    this.el?.focus();
   }
 
   @Method() async toggle(): Promise<void> {
@@ -346,12 +358,13 @@ export class CalcitePopover {
 
     return closeButton ? (
       <button
+        ref={closeButtonEl => (this.closeButtonEl = closeButtonEl)}
         aria-label={textClose}
         title={textClose}
         class={{ [CSS.closeButton]: true }}
         onClick={this.hide}
       >
-        <CalciteIcon size="16" path={x16} />
+        <calcite-icon icon="x" scale="s"></calcite-icon>
       </button>
     ) : null;
   }