+++ stencil.config.ts
@@ -43,7 +43,9 @@ export const config: Config = {
       ]
     },
     { components: ["calcite-tooltip"] },
-    { components: ["calcite-tree", "calcite-tree-item"] }
+    { components: ["calcite-tree", "calcite-tree-item"] },
+    { components: ["calcite-card"] },
+    { components: ["calcite-icon"] }
   ],
   outputTargets: [
     { type: "dist-hydrate-script" },